review(halo2_ivc): address team's comments on test infrastructure
