+++ mithril-stm/src/circuits/halo2_ivc/tests/transitions/negative/genesis.rs
@@ -1,0 +1,124 @@
+use super::*;
+
+use crate::circuits::halo2_ivc::{
+    state::State,
+    tests::common::{
+        asset_readers::load_embedded_genesis_step_output_asset,
+        generators::{
+            GENESIS_EPOCH, build_asset_generation_setup, build_genesis_base_case_next_state,
+            build_genesis_base_case_witness,
+        },
+        helpers::{
+            assert_recursive_mock_prover_rejects, build_mock_prover_public_inputs,
+            build_recursive_mock_prover_setup, build_trivial_mock_prover_circuit,
+        },
+    },
+};
+
+#[test]
+fn merkle_root_tampered_is_rejected() {
+    // Asset-based check: circuit enforces merkle_root = 0 at the genesis base case.
+    assert_step_output_rejects_tampered_state(
+        load_embedded_genesis_step_output_asset,
+        "genesis step output",
+        |s| s.merkle_root = F::ONE,
+        "proof with tampered merkle_root should be rejected by the verifier",
+    );
+}
+
+#[test]
+fn protocol_params_tampered_is_rejected() {
+    // Asset-based check: circuit enforces protocol_params = 0 at the genesis base case.
+    assert_step_output_rejects_tampered_state(
+        load_embedded_genesis_step_output_asset,
+        "genesis step output",
+        |s| s.protocol_params = F::ONE,
+        "proof with tampered protocol_params should be rejected by the verifier",
+    );
+}
+
+#[test]
+fn next_merkle_root_tampered_is_rejected() {
+    // Asset-based check: circuit enforces next_merkle_root is extracted from the genesis message preimage.
+    assert_step_output_rejects_tampered_state(
+        load_embedded_genesis_step_output_asset,
+        "genesis step output",
+        |s| s.next_merkle_root = F::ONE,
+        "proof with tampered next_merkle_root should be rejected by the verifier",
+    );
+}
+
+#[test]
+fn next_protocol_params_tampered_is_rejected() {
+    // Asset-based check: circuit enforces next_protocol_params is extracted from the genesis message preimage.
+    assert_step_output_rejects_tampered_state(
+        load_embedded_genesis_step_output_asset,
+        "genesis step output",
+        |s| s.next_protocol_params = F::ONE,
+        "proof with tampered next_protocol_params should be rejected by the verifier",
+    );
+}
+
+#[test]
+fn counter_tampered_is_rejected() {
+    // Asset-based check: circuit enforces counter transitions 0 → 1 at the genesis base case.
+    assert_step_output_rejects_tampered_state(
+        load_embedded_genesis_step_output_asset,
+        "genesis step output",
+        |s| s.counter = F::from(2u64),
+        "proof with tampered counter should be rejected by the verifier",
+    );
+}
+
+#[test]
+fn current_epoch_tampered_is_rejected() {
+    // Asset-based check: circuit enforces current_epoch is extracted from the genesis message preimage.
+    assert_step_output_rejects_tampered_state(
+        load_embedded_genesis_step_output_asset,
+        "genesis step output",
+        |s| s.current_epoch = F::ONE,
+        "proof with tampered current_epoch should be rejected by the verifier",
+    );
+}
+
+#[test]
+fn msg_tampered_is_rejected() {
+    // Asset-based check: circuit enforces msg equals the genesis message committed in the Global public inputs.
+    assert_step_output_rejects_tampered_state(
+        load_embedded_genesis_step_output_asset,
+        "genesis step output",
+        |s| s.msg = F::ONE,
+        "proof with tampered msg should be rejected by the verifier",
+    );
+}
+
+mod slow {
+    use super::*;
+
+    /// Builds a genesis circuit with a tampered next state and asserts the MockProver rejects it.
+    ///
+    /// `tamper` receives the correct genesis next state before it is passed to the
+    /// public inputs, allowing each test to corrupt exactly the field it wants to verify.
+    fn assert_genesis_circuit_rejects_tampered_next_state(tamper: impl FnOnce(&mut State)) {
+        let setup = build_asset_generation_setup();
+        let mock_prover_setup = build_recursive_mock_prover_setup(&setup);
+
+        let witness = build_genesis_base_case_witness(&setup);
+
+        let mut tampered_state = build_genesis_base_case_next_state(&setup, GENESIS_EPOCH);
+        tamper(&mut tampered_state);
+
+        let circuit =
+            build_trivial_mock_prover_circuit(&mock_prover_setup, State::genesis(), witness);
+        let public_inputs = build_mock_prover_public_inputs(&mock_prover_setup, &tampered_state);
+
+        assert_recursive_mock_prover_rejects(circuit, public_inputs);
+    }
+
+    #[test]
+    fn circuit_rejects_msg_inconsistent_with_preimage() {
+        // MockProver check that the in-circuit Blake2b hash constraint between
+        // msg_preimage bytes and the resulting msg field is wired correctly.
+        assert_genesis_circuit_rejects_tampered_next_state(|s| s.msg = F::ONE);
+    }
+}
